requirements.txt
pandas
lxml
streamlit
matplotlib
mysql-connector-python
